rsvp/pyproject.toml
[project]
name = "rsvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "plotly>=6.6.0",
    "pyqt6>=6.11.0",
    "scipy>=1.17.1",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
]
